feat(25-testing/1): Resolve test framework compatibility issues
Rework example to make it fit into our automated test framework that
relies on '.exercise.go' files too.
