+++ 25-testing/01-testify/.README.md
@@ -4,6 +4,9 @@ In this task, you need to test the function `{{ index . "name" }}` in Go using *
 
 ## Function Description
 
+The implementation is available in [exercise_test.go](exercise_test.go), if needed.
+Below we show some examples of the expected behaviors.
+
 {{- if eq (index . "name") "strstr" }}
 
 ### `strstr(str string, substr string) int`
@@ -72,21 +75,21 @@ strncpy(&dest, "", 1)      // dest == ""
 
 {{- if eq (index . "name") "strstr" }}
 
-- `StrStrMatch` — test when the substring exists
-- `StrStrNoMatch` — test when the substring does not exist
-- `StrStrEmptySubString` — test when the substring is empty
+- `StrStrMatch(t *testing.T)` — test when the substring exists
+- `StrStrNoMatch(t *testing.T)` — test when the substring does not exist
+- `StrStrEmptySubString(t *testing.T)` — test when the substring is empty
 {{- end }}
 {{- if eq (index . "name") "strncat" }}
 
-- `StrNCatInbounds` — append a part of the source within index bounds
-- `StrNCatOutOfBounds` — append more than the source length
-- `StrNCatEmptySource` — append from an empty source
+- `StrNCatInbounds(t *testing.T)` — append a part of the source within index bounds
+- `StrNCatOutOfBounds(t *testing.T)` — append more than the source length
+- `StrNCatEmptySource(t *testing.T)` — append from an empty source
 {{- end }}
 {{- if eq (index . "name") "strncpy" }}
 
-- `StrNCpyINbounds` — copy part of the source within index bounds
-- `StrNCpyEmptyDestination` — copy into an empty destination
-- `StrNCpyEmptySource` — copy from an empty source
+- `StrNCpyINbounds(t *testing.T)` — copy part of the source within index bounds
+- `StrNCpyEmptyDestination(t *testing.T)` — copy into an empty destination
+- `StrNCpyEmptySource(t *testing.T)` — copy from an empty source
 {{- end }}
 
 Insert the code into the file `exercise.go` at the placeholder `// INSERT YOUR CODE HERE`.
