+++ 25-testing/01-testify/exercise.go
@@ -1,4 +1,13 @@
 package testimony
 
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
 // DO NOT REMOVE THIS COMMENT
 //go:generate go run ../../exercises-cli.go -student-id=$STUDENT_ID generate
+
+
+// INSERT YOUR CODE HERE
